# openapi.yaml
openapi: 3.0.0
servers:
- url: https://forge.api.prod-preview.openshift.io/creator/
  description: Development server
- url: https://forge.api.openshift.io/creator/
  description: Production server
- url: http://localhost:8081
  description: Local Development
- url: http://localhost:8080
  description: Local Dist

info:
  version: "1.0.0"
  title: Launcher Creator Backend API
  description: >-
    This document defines the REST endpoints exposed by the
    launcher-creator-backend component

security:
- application:
  - read
  - write

paths:
  /enums:
    get:
      summary: Returns all available enumeration definitions
      security: []
      description: >-
        This endpoint returns all known enumerations
      tags:
      - catalog
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enumerations'
        '404':
          description: Not Found
  /enums/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      summary: Returns the selected enumeration definition
      security: []
      description: >-
        This endpoint returns a given enumeration
      tags:
      - catalog
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enumeration'
        '404':
          description: Not Found
  /capabilities:
    get:
      summary: Returns the available Capabilities
      security: []
      description: >-
        This endpoint returns all known Capabilities
      tags:
      - catalog
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Capabilities'
        '404':
          description: Not Found
  /zip:
    post:
      summary: Create ZIP containing a generated project and returns and id to download it
      security: []
      description: >-
        This endpoint returns and id to download the ZIP containing code that was
        generated using the specified Capabilities
      tags:
      - project
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZipRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    description: The id to download the zip
                    type: string
        '404':
          description: Not Found
  /download:
    get:
      summary: Returns a ZIP containing a generated project for the specified id
      security: []
      description: >-
        This endpoint returns a ZIP containing a generated project for the specified id
      tags:
      - project
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: string
        description: The id returned by the zip method
      responses:
        '200':
          description: OK
          content:
            application/zip: {}
        '404':
          description: Not Found
  /launch:
    post:
      summary: Pushes a generated project to GitHub and OpenShift
      security:
      - bearerAuth: []
      description: >-
        This endpoint pushes code that was generated using the specified
        Capabilities to a user's GitHub and OpenShift accounts
      tags:
      - project
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LaunchRequest'
      responses:
        '200':
          description: OK
        '404':
          description: Not Found

components:
  schemas:
    Enumeration:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        metadata:
          type: object

    Enumerations:
      type: array
      items:
        $ref: '#/components/schemas/Enumeration'

    Capabilities:
      type: array
      items:
        type: object
        properties:
          module:
            type: string
          name:
            type: string
          description:
            type: string
          props:
            type: array
            items:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                description:
                  type: string
                type:
                  type: string
                required:
                  type: boolean

    ZipRequest:
      type: object
      properties:
        project:
          $ref: '#/components/schemas/Application'
      required:
      - project

    LaunchRequest:
      type: object
      properties:
        project:
          $ref: '#/components/schemas/Application'
        clusterId:
          description: The OpenShift cluster to use
          type: string
        projectName:
          description: The OpenShift project name to be used
          type: string
        gitOrganization:
          description: The Git Organization to push the code
          type: string
        gitRepository:
          description: The Git Repository to push the code
          type: string
      required:
      - project
      - clusterId
      - projectName
      - gitOrganization
      - gitRepository

    Application:
      type: object
      properties:
        application:
          description: The application name to use for the project
          type: string
        tiers:
          description: The application tiers to use for the project
          type: array
          items:
              type: object
              properties:
                tier:
                  description: The name of the application subFolderName
                  type: string
                shared:
                  description: The shared properties that get applied to each selected Capability
                  type: object
                  properties:
                    runtime:
                      description: The runtime to use for the project
                      type: object
                      properties:
                        name:
                          description: The name of the runtime
                          type: string
                        version:
                          description: The version of the runtime
                          type: string
                      required:
                      - name
                capabilities:
                  description: The capabilities to add to the project
                  type: array
                  items:
                    type: object
                    properties:
                      module:
                        description: The id of the selected Capability
                        type: string
                      props:
                        description: The properties for the selected Capability
                        type: object
                    required:
                    - module
              required:
              - capabilities
      required:
      - application
      - tiers

  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes